+++ lib/src/git_backend.rs
@@ -586,20 +586,34 @@ fn gix_open_opts_from_settings(settings: &UserSettings) -> gix::open::Options {
 }
 
 /// Parses the `jj:conflict-labels` header value if present.
-fn extract_conflict_labels_from_commit(commit: &gix::objs::CommitRef) -> Merge<String> {
+fn extract_conflict_labels_from_commit(commit: &gix::objs::CommitRef) -> Result<Merge<String>, ()> {
     let Some(value) = commit
         .extra_headers()
         .find(JJ_CONFLICT_LABELS_COMMIT_HEADER)
     else {
-        return Merge::resolved(String::new());
+        return Ok(Merge::resolved(String::new()));
     };
 
-    str::from_utf8(value)
-        .expect("labels should be valid utf8")
+    let labels = str::from_utf8(value)
+        .map_err(|_| ())?
         .split_terminator('\n')
         .map(str::to_owned)
-        .collect::<MergeBuilder<_>>()
-        .build()
+        .collect_vec();
+    if labels.len() == 1 || labels.len() % 2 == 0 {
+        return Err(());
+    }
+    Ok(Merge::from_vec(labels))
+}
+
+fn validate_conflict_labels(
+    root_tree: &Merge<TreeId>,
+    conflict_labels: &Merge<String>,
+) -> Result<(), ()> {
+    if root_tree.num_sides() == conflict_labels.num_sides() {
+        Ok(())
+    } else {
+        Err(())
+    }
 }
 
 /// Parses the `jj:trees` header value if present, otherwise returns the
@@ -656,13 +670,22 @@ fn commit_from_git_without_root_parent(
     };
     // If the commit is a conflict, the conflict labels are stored in a commit
     // header separately from the trees.
-    let conflict_labels = extract_conflict_labels_from_commit(&commit);
+    let conflict_labels = extract_conflict_labels_from_commit(&commit)
+        .map_err(|()| to_read_object_err("Invalid jj:conflict-labels header", id))?;
     // Conflicted commits written before we started using the `jj:trees` header
     // (~March 2024) may have the root trees stored in the extra metadata table
     // instead. For such commits, we'll update the root tree later when we read the
     // extra metadata.
     let root_tree = extract_root_tree_from_commit(&commit)
         .map_err(|()| to_read_object_err("Invalid jj:trees header", id))?;
+    if commit
+        .extra_headers()
+        .find(JJ_TREES_COMMIT_HEADER)
+        .is_some()
+    {
+        validate_conflict_labels(&root_tree, &conflict_labels)
+            .map_err(|()| to_read_object_err("Invalid jj:conflict-labels header", id))?;
+    }
     // Use lossy conversion as commit message with "mojibake" is still better than
     // nothing.
     // TODO: what should we do with commit.encoding?
@@ -1265,6 +1288,8 @@ impl Backend for GitBackend {
             let extras = table.get_value(id.as_bytes()).unwrap();
             deserialize_extras(&mut commit, extras);
         }
+        validate_conflict_labels(&commit.root_tree, &commit.conflict_labels)
+            .map_err(|()| to_read_object_err("Invalid jj:conflict-labels header", id))?;
         Ok(commit)
     }